Module uvls::core

This Module Represents the UVL Language Server (LSP) Logic

Re-exports

pub use ast::*;

pub use cache::*;

pub use check::*;

pub use config::*;

pub use document::*;

pub use module::*;

pub use parse::*;

pub use pipeline::*;

pub use semantic::*;

pub use util::*;

Modules

ast

Easy to work with AST parsing and util.

cache

Find files control what files need to be relinked

check

config

Configuration is stored in json files like this

document

File update logic

module

parse

green tree parsing (tree-sitter)

pipeline

multi-stage compiler, it provides access to files in various compilation stages. Constists of multiple subsystems like: green tree parsing, linking, error reporting, smt checking etc.

query

resolve

Path resolving on file level

semantic

Contains a central struct RootGraph. This struct includes the latest linked uvl and config files, it can be thought of as a snapshot of all related documents. It can be accessed through the pipeline.

util

This Module provides basic utility functions for the LSP and SMTSolver (mainly byte offsets und byte ranges)

Macros

info

Logs a message at the info level.