Struct uvls::core::ast::Ast

struct Ast {
    keywords: Vec<Keyword>,
    namespace: Option<Path>,
    includes: Vec<LanguageLevelDecl>,
    import: Vec<Import>,
    features: Vec<Feature>,
    constraints: Vec<ConstraintDecl>,
    attributes: Vec<Attribute>,
    references: Vec<Reference>,
    groups: Vec<Group>,
    dirs: Vec<Dir>,
    structure: TreeMap,
    index: HashMap<(Symbol, Ustr, SymbolKind), Symbol>,
}

Ast container each symbol kind lives in its own vector

Fields
keywords: Vec<Keyword>namespace: Option<Path>includes: Vec<LanguageLevelDecl>import: Vec<Import>features: Vec<Feature>constraints: Vec<ConstraintDecl>attributes: Vec<Attribute>references: Vec<Reference>groups: Vec<Group>dirs: Vec<Dir>structure: TreeMapindex: HashMap<(Symbol, Ustr, SymbolKind), Symbol>
Implementations

impl Ast

pub fn import_prefix(&self, sym: Symbol) -> &[Ustr]

fn lookup<F: FnMut(Symbol)>(&self, sym: Symbol, prefix: Ustr, f: F)

call f for each child under sym and prefix

fn name(&self, sym: Symbol) -> Option<Ustr>

fn lsp_range(&self, sym: Symbol, source: &Rope) -> Option<Range>

fn span(&self, sym: Symbol) -> Option<Range<usize>>

source range for a symbol if there is any

fn children( &self, sym: Symbol, ) -> impl Iterator<Item = Symbol> + DoubleEndedIterator + '_

fn all_imports(&self) -> impl Iterator<Item = Symbol> + DoubleEndedIterator

utility iterators over different elements of interest

fn get_import(&self, index: usize) -> Option<&Import>

fn all_features(&self) -> impl Iterator<Item = Symbol>

fn get_feature(&self, index: usize) -> Option<&Feature>

fn containe_feature(&self, name: Ustr) -> bool

fn get_attribute(&self, index: usize) -> Option<&Attribute>

fn all_attributes(&self) -> impl Iterator<Item = Symbol>

fn containe_attribute(&self, name: Ustr) -> bool

fn all_references(&self) -> impl Iterator<Item = Symbol>

fn all_constraints(&self) -> impl Iterator<Item = Symbol>

fn all_lang_lvls(&self) -> impl Iterator<Item = Symbol>

fn find(&self, offset: usize) -> Option<Symbol>

Search a symbol by byte offset in O(N)

fn find_all_of(&self, current: Symbol, name: Ustr) -> Vec<Symbol>