Struct uvls::core::ast::AstDocument

pub struct AstDocument {
    ast: Ast,
    pub source: Rope,
    pub tree: Tree,
    pub timestamp: Instant,
    pub errors: Vec<ErrorInfo>,
    pub path: Vec<Ustr>,
    pub uri: Url,
    pub id: FileID,
}

Combines the AST with metadata, this is also a public interface to the AST.

Fields
ast: Astsource: Ropetree: Treetimestamp: Instanterrors: Vec<ErrorInfo>path: Vec<Ustr>uri: Urlid: FileID
Implementations

impl AstDocument

pub fn parent(&self, sym: Symbol, merge_root_features: bool) -> Option<Symbol>

pub fn scope(&self, sym: Symbol) -> Symbol

parent feature of an attribute

pub fn name(&self, sym: Symbol) -> Option<Ustr>

pub fn all_lang_lvls(&self) -> impl Iterator<Item = Symbol>

iterators over diffrent symbole types in the tree

pub fn all_imports(&self) -> impl Iterator<Item = Symbol> + DoubleEndedIterator

pub fn get_import(&self, index: usize) -> Option<&Import>

pub fn all_features(&self) -> impl Iterator<Item = Symbol>

pub fn get_feature(&self, index: usize) -> Option<&Feature>

pub fn containe_feature(&self, name: Ustr) -> bool

pub fn get_attribute(&self, index: usize) -> Option<&Attribute>

pub fn all_attributes(&self) -> impl Iterator<Item = Symbol>

pub fn containe_attribute(&self, name: Ustr) -> bool

pub fn all_references(&self) -> impl Iterator<Item = Symbol>

pub fn all_constraints(&self) -> impl Iterator<Item = Symbol>

pub fn lang_lvl(&self, sym: Symbol) -> Option<&LanguageLevel>

pub fn group_mode(&self, sym: Symbol) -> Option<GroupMode>

pub fn constraint(&self, sym: Symbol) -> Option<&ConstraintDecl>

pub fn constraints(&self) -> &[ConstraintDecl]

pub fn imports(&self) -> &[Import]

pub fn value(&self, sym: Symbol) -> Option<&Value>

pub fn direct_children( &self, sym: Symbol, ) -> impl Iterator<Item = Symbol> + DoubleEndedIterator + '_

pub fn get_reference(&self, index: usize) -> Option<&Reference>

pub fn lsp_range(&self, sym: Symbol) -> Option<Range>

pub fn span(&self, sym: Symbol) -> Option<Range<usize>>

pub fn namespace(&self) -> Option<&Path>

pub fn path(&self, sym: Symbol) -> &[Ustr]

pub fn import_prefix(&self, sym: Symbol) -> &[Ustr]

pub fn depth(&self, sym: Symbol) -> u32

pub fn lookup<'a, F: Fn(Symbol) -> bool + 'a>( &'a self, root: Symbol, path: &'a [Ustr], filter: F, ) -> impl Iterator<Item = Symbol> + 'a

Find all symboles under root with prefix path. Search branches can be aborted with a filter

pub fn lookup_import<'a>( &'a self, path: &'a [Ustr], ) -> impl Iterator<Item = (Symbol, &'a [Ustr])>

Find imports for any prefix of path

pub fn lookup_with_binding<'a, F: Fn(Symbol) -> bool + 'a>( &'a self, root: Symbol, path: &'a [Ustr], filter: F, ) -> impl Iterator<Item = Vec<Symbol>> + 'a

Also track the binding for path, each segment of path has some symbole bound to it

pub fn get_all_entities(&self, path: &[Ustr]) -> Vec<Symbol>

returns all Symbols with same name, used for cardinality resolving

pub fn get_symbols(&self, path: Ustr) -> Vec<Symbol>

pub fn prefix(&self, sym: Symbol) -> Vec<Ustr>

prefix of sym from root

pub fn type_of(&self, sym: Symbol) -> Option<Type>

pub fn find(&self, offset: usize) -> Option<Symbol>

pub fn visit_named_children<F: FnMut(Symbol, &[Ustr]) -> bool>( &self, root: Symbol, merge_root_features: bool, f: F, )

All children under root, when merge_root_features sub features are ignored

pub fn visit_named_children_depth<F: FnMut(Symbol, &[Ustr], usize) -> bool>( &self, root: Symbol, merge_root_features: bool, f: F, )

Iterate all named symbole under root and track their prefix from root

pub fn visit_children<F: FnMut(Symbol) -> bool>( &self, root: Symbol, merge_root_features: bool, f: F, )

pub fn find_all_of(&self, name: Ustr) -> Vec<Symbol>

pub fn visit_children_depth<F: FnMut(Symbol, u32) -> bool>( &self, root: Symbol, merge_root_features: bool, f: F, )

Iterate all named symbole under root

pub fn visit_attributes<'a, F: FnMut(Symbol, Symbol, &[Ustr])>( &self, root: Symbol, f: F, )

Iterate all attributes under root, and track theire associated feature

pub fn new(source: Rope, tree: Tree, uri: Url, timestamp: Instant) -> Self