Module uvls::core::ast::transform

Transform a tree-sitter tree into the Ast ECS via recursive decent While parsing we keep a mutable state to store entities and errors

Structs

VisitorState 🔒

Functions

check_langlvls 🔒

check_no_extra_blk 🔒

report error if a node has cardinality or attributes

check_simple_blk 🔒

Report error if a node has children,cardinality or attributes

opt_aggreate_op 🔒

opt_aggregate 🔒

opt_arithmetic_minor 🔒

opt_attrib_expr 🔒

opt_bool 🔒

opt_boolean_minor 🔒

opt_cardinality 🔒

opt_constraint 🔒

opt_equation 🔒

opt_function_args 🔒

opt_int 🔒

opt_integer 🔒

opt_integer_op 🔒

opt_lang_lvl 🔒

opt_logic_op 🔒

opt_major_lang_lvl 🔒

opt_name 🔒

opt_number 🔒

opt_numeric 🔒

opt_numeric_op 🔒

opt_path 🔒

opt_string 🔒

opt_type_minor 🔒

opt_value 🔒

visit_attribute_value 🔒

visit_attributes 🔒

visit_blk_decl 🔒

visit_constraint 🔒

visit_constraint_decl 🔒

visit_constraint_list 🔒

visit_constraints 🔒

visit_feature 🔒

visit_features 🔒

visit_group 🔒

visit_import 🔒

visit_import_decl 🔒

visit_include 🔒

visit_lang_lvl 🔒

visit_namespace 🔒

parse a namespace

visit_ref 🔒

visit_root

visits all valid children of a tree-sitter (green tree) recursively to translate them into the AST(red tree)

visit_top_lvl 🔒