Module uvls::core::config

Configuration is stored in json files like this

{
    "file":"file.uvl",
    "config":{
          "submodels.subfile":{
              "subfeature":true
          }
          "someFeatures":true,
          "someOtherFeature":1.0,
          "someFeature.attribute1":"test"
    }

}

This representation is very compact since it avoids rewriting long import prefixes but slightly more complex than just using the direct raw path to external symbols. JSON parsing is done with tree-sitter and not serde because there currently is no solid serde json crate for span information and partial parsing so error reporting becomes impossible.

Structs

ConfigDocument

FileConfig

This defines a config for the corresponding file

State 🔒

Enums

CardinalityEntry

This enum is used for storing a cardinality of a feature

ConfigEntry

This Enum is used for Serialization and Deserialization of a Config

ConfigValue

This Enum stores the value configured by the user or loaded from a json config File.

JSONItem 🔒

Traits

Overlaps 🔒

Functions

completion_query

This function helps define the context for a completion.

estimate_env_json

estimate_json_item 🔒

find_file_id

find_json_key 🔒

Try to extract the json value under key from node,

find_selected_json_key 🔒

json_path 🔒

find the path of the object containing node(ignores arrays)

offset 🔒

opt_configs 🔒

Prase a configuration object

parse_json

parse_json_key 🔒

selected_json_object 🔒

select the nearest object containing pos

visit_file 🔒

visit_root 🔒