Struct uvls::core::module::ConfigModule

pub struct ConfigModule {
    pub module: Arc<Module>,
    pub values: HashMap<ModuleSymbol, ConfigValue>,
    pub source_map: HashMap<ModuleSymbol, Range<usize>>,
}

Configuration with a module and resolved configuration symboles

Fields
module: Arc<Module>values: HashMap<ModuleSymbol, ConfigValue>source_map: HashMap<ModuleSymbol, Range<usize>>
Implementations

impl ConfigModule

fn serialize_rec(&self, path: &[Ustr], i: InstanceID) -> ConfigEntry

fn serialize_rec_file( &self, sym: Symbol, file: &AstDocument, i: InstanceID, ) -> Vec<ConfigEntry>

serialize file recursive while accommodate for cardinality

pub fn serialize(&self) -> Vec<ConfigEntry>

Turns a the set of linear configuration values of this module into theire recusive from used in json