Module uvls::ide::completion

Handle Completion and code suggestions for IDEs

All things completion related happen in here, the process is roughly as follows:

Find the current context using the latest draft and editor position

Find good completions in this context

The completion context includes:

Meta information on the cursor position eg. Are we currently in a path or an empty line etc.

The semantic context eg. do we need a constraint or a number

A optional path prefix and suffix. The suffix is used as a weight for completions using the jaro winkler distance, the prefix is a filter restricting possible completions.

To weigh completions we use a simple weight function with hand picked weights for parameters like length or type correctness

Structs

CompletionOpt 🔒

CompletionQuery

ModulePath 🔒

TopN 🔒

Enums

CompletionEnv

CompletionFormatter

CompletionKind 🔒

CompletionOffset

Section

TextOP 🔒

Statics

AVG_WEIGHT_THRESHOLD 🔒

MAX_N 🔒

All things completion related happen in here, the process is roughly as follows:

MIN_WEIGHT 🔒

W_LEN 🔒

W_TYPE 🔒

Functions

add_function_keywords 🔒

add_group_keywords 🔒

add_keywords 🔒

add_lang_lvl_arithmetic 🔒

add_lang_lvl_boolean 🔒

add_lang_lvl_major_keywords 🔒

add_lang_lvl_type 🔒

add_logic_op 🔒

add_numeric_op 🔒

add_string_op 🔒

add_top_lvl_keywords 🔒

completion_symbol 🔒

completion_symbol_local 🔒

completion_weight 🔒

compute_completions

compute_completions_impl 🔒

compute_constraint_completion 🔒

contains

encode_float 🔒

estimate_context 🔒

estimate_env 🔒

estimate_expr

find_section

longest_path

make_path

make_relativ_path 🔒

node_at 🔒

path_len 🔒

position_to_node 🔒

resolve_name 🔒

starts_with