Module uvls::webview::frontend

This Module is used for frontend rendering

It provides multiple helper functions which convert into Elements to display

Structs

ConfigInputProps ğŸ”’

FileEntryProps ğŸ”’

IconProps ğŸ”’

RealInputProps ğŸ”’

ValueProps ğŸ”’

Functions

App

ConfigInput ğŸ”’

FileEntry ğŸ”’

Props

Icon ğŸ”’

returns the requested Icon Element to display in the frontend

RealInput ğŸ”’

Props

Spinner ğŸ”’

Render UVL content to HTML

Value ğŸ”’

Props

file_values_iter ğŸ”’

icon ğŸ”’

to_number ğŸ”’