Module uvls::core::ast::transform

Transform a tree-sitter tree into the Ast ECS via recursive decent While parsing we keep a mutable state to store entities and errors

Structs

VisitorState ğŸ”’

Functions

check_langlvls ğŸ”’

check_no_extra_blk ğŸ”’

report error if a node has cardinality or attributes

check_simple_blk ğŸ”’

Report error if a node has children,cardinality or attributes

opt_aggreate_op ğŸ”’

opt_aggregate ğŸ”’

opt_arithmetic_minor ğŸ”’

opt_attrib_expr ğŸ”’

opt_bool ğŸ”’

opt_boolean_minor ğŸ”’

opt_cardinality ğŸ”’

opt_constraint ğŸ”’

opt_equation ğŸ”’

opt_function_args ğŸ”’

opt_int ğŸ”’

opt_integer ğŸ”’

opt_integer_op ğŸ”’

opt_lang_lvl ğŸ”’

opt_logic_op ğŸ”’

opt_major_lang_lvl ğŸ”’

opt_name ğŸ”’

opt_number ğŸ”’

opt_numeric ğŸ”’

opt_numeric_op ğŸ”’

opt_path ğŸ”’

opt_string ğŸ”’

opt_type_minor ğŸ”’

opt_value ğŸ”’

visit_attribute_value ğŸ”’

visit_attributes ğŸ”’

visit_blk_decl ğŸ”’

visit_constraint ğŸ”’

visit_constraint_decl ğŸ”’

visit_constraint_list ğŸ”’

visit_constraints ğŸ”’

visit_feature ğŸ”’

visit_features ğŸ”’

visit_group ğŸ”’

visit_import ğŸ”’

visit_import_decl ğŸ”’

visit_include ğŸ”’

visit_lang_lvl ğŸ”’

visit_namespace ğŸ”’

parse a namespace

visit_ref ğŸ”’

visit_root

visits all valid children of a tree-sitter (green tree) recursively to translate them into the AST(red tree)

visit_top_lvl ğŸ”’