Struct uvls::core::pipeline::AsyncPipeline

pub struct AsyncPipeline {
    drafts: Arc<DashMap<Url, DraftState>>,
    tx_link: Sender<LinkMsg>,
    tx_err: Sender<DiagnosticUpdate>,
    rx_root: Receiver<Arc<RootGraph>>,
    tx_dirty_tree: Sender<()>,
    revision_counter: Arc<AtomicU64>,
    client: Client,
    inlay_handler: InlayHandler,
}

All the parsing components and their consumers in a central interface

Fields
drafts: Arc<DashMap<Url, DraftState>>tx_link: Sender<LinkMsg>tx_err: Sender<DiagnosticUpdate>rx_root: Receiver<Arc<RootGraph>>tx_dirty_tree: Sender<()>revision_counter: Arc<AtomicU64>client: Clientinlay_handler: InlayHandler
Implementations

impl AsyncPipeline

pub fn new(client: Client) -> Self

pub fn touch(&self, uri: &Url)

pub fn inlay_state(&self) -> &InlayHandler

pub fn client(&self) -> Client

pub fn subscribe_dirty_tree(&self) -> Receiver<()>

pub fn open(&self, uri: Url, text: String, state: DocumentState)

pub fn should_load(&self, uri: &Url, time: SystemTime) -> bool

pub fn stat(&self, uri: &Url) -> Option<(Instant, DocumentState)>

pub async fn delete(&self, uri: &Url, state: DocumentState)

pub fn update(&self, params: DidChangeTextDocumentParams)

pub async fn snapshot_draft( &self, uri: &Url, ) -> Result<Option<Draft>, Box<dyn Error + Send + Sync>>

Wait for latest draft version

pub async fn snapshot_root( &self, uri: &Url, ) -> Result<Arc<RootGraph>, Box<dyn Error + Send + Sync>>

Wait until uri is found in the linked root graph

pub fn root(&self) -> Receiver<Arc<RootGraph>>

pub async fn snapshot_root_sync( &self, uri: &Url, timestamp: Instant, ) -> Result<Arc<RootGraph>, Box<dyn Error + Send + Sync>>

wait until uri newer than timestamp in the root graph

pub async fn sync_root<F: Fn(&RootGraph) -> bool>( &self, f: F, ) -> Result<Arc<RootGraph>, Box<dyn Error + Send + Sync>>

pub async fn sync_root_global( &self, ) -> Result<Arc<RootGraph>, Box<dyn Error + Send + Sync>>

wait until ALL parsing is done and root is clean

pub async fn snapshot( &self, uri: &Url, sync: bool, ) -> Result<Option<(Draft, Arc<RootGraph>)>, Box<dyn Error + Send + Sync>>

get latest draft and sync root