Module uvls::core::resolve

Path resolving on file level

Imports between files allow for multiple possible meanings hence depth first search is required This also containes functions for attribute aggregates and path-symbol binding and type checking

resolve_file: Type resolving for an AstDocument

resolve_*: nameresultion under various conditions.

Structs

TypeResolveContext 🔒

Enums

ResolveState

Traits

AstContainer

Functions

commit_expr 🔒

Fix types

common_prefix

estimate_types

Best effort type resolve for a single tree-sitter expression

gather_expr_options 🔒

Find possible types

resolve

Find all symboles from origin under path

resolve_attributes

resolve_attributes_with_feature

Find all attributes in orgin under context, allso gives the prefix for each attribut

resolve_constraint 🔒

Since there can be multiple possible interpretations, two pases are used for equations. First pass: Gather a set of all possible types for each equation side. Second pass: Pick a common type and resolve each expression with the choosen type We could simply forbid alias import and feature names but thats boring.

resolve_file

Best effort tupe resolving

resolve_with_bind

Find all symboles from origin under path while keeping track of what sections path are bound to what symbol

select_type 🔒

Type Aliases

RefMap 🔒