Module uvls::ide::completion

Handle Completion and code suggestions for IDEs

All things completion related happen in here, the process is roughly as follows:

Find the current context using the latest draft and editor position

Find good completions in this context

The completion context includes:

Meta information on the cursor position eg. Are we currently in a path or an empty line etc.

The semantic context eg. do we need a constraint or a number

A optional path prefix and suffix. The suffix is used as a weight for completions using the jaro winkler distance, the prefix is a filter restricting possible completions.

To weigh completions we use a simple weight function with hand picked weights for parameters like length or type correctness

Structs

CompletionOpt ğŸ”’

CompletionQuery

ModulePath ğŸ”’

TopN ğŸ”’

Enums

CompletionEnv

CompletionFormatter

CompletionKind ğŸ”’

CompletionOffset

Section

TextOP ğŸ”’

Statics

AVG_WEIGHT_THRESHOLD ğŸ”’

MAX_N ğŸ”’

All things completion related happen in here, the process is roughly as follows:

MIN_WEIGHT ğŸ”’

W_LEN ğŸ”’

W_TYPE ğŸ”’

Functions

add_function_keywords ğŸ”’

add_group_keywords ğŸ”’

add_keywords ğŸ”’

add_lang_lvl_arithmetic ğŸ”’

add_lang_lvl_boolean ğŸ”’

add_lang_lvl_major_keywords ğŸ”’

add_lang_lvl_type ğŸ”’

add_logic_op ğŸ”’

add_numeric_op ğŸ”’

add_string_op ğŸ”’

add_top_lvl_keywords ğŸ”’

completion_symbol ğŸ”’

completion_symbol_local ğŸ”’

completion_weight ğŸ”’

compute_completions

compute_completions_impl ğŸ”’

compute_constraint_completion ğŸ”’

contains

encode_float ğŸ”’

estimate_context ğŸ”’

estimate_env ğŸ”’

estimate_expr

find_section

longest_path

make_path

make_relativ_path ğŸ”’

node_at ğŸ”’

path_len ğŸ”’

position_to_node ğŸ”’

resolve_name ğŸ”’

starts_with